vents (physics and particles)
i am actually really happy with this. the vents can be rotated in any direction and will work according to the rotation, both the physics and the particles. I can also change the length of them  through an exported value and the particle values automatically change to match, as well as the collision shape :)

I havent made a dedicated model for it yet but I made it a texture for the mean time